assets/portfolio/pesquise-o-preco/pesquiseopreco/requirements.txt
Django==1.9
django-excel==0.0.10
django-fontawesome==0.3.1
Jinja2==2.10
lml==0.0.1
MarkupSafe==1.0
Pillow==5.0.0
pyexcel==0.5.7
pyexcel-handsontable==0.0.1
pyexcel-io==0.5.6
pyexcel-webio==0.1.4
pyexcel-xls==0.5.5
PyYAML==3.12
texttable==1.2.1
xlrd==1.1.0
xlwt==1.3.0
